Move out tests of a deprecated module to work around `#[test]` bugs
https://github.com/rust-lang/rust/issues/47238
